social_media_api/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
Pillow==10.3.0
django-filter==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
